// src/app/types/CountryInfo.ts
export interface CountryInfo {
  name: string;
  capital: string;
  continent: string;
  population: number;
  flag: string;
  map: string;
  languages: string;
  history: string;
  upcomingHolidays: object;
  subregion: string;
  coa: string;
  area: number;
  areaComparative: string;
  climate: string;
  shortDescription: string;
  longLat: string;
  etymology: string;
  bbd: number;
  cad: number;
  eur: number;
  usd: number;
  lastUpdated: number;
}